refactor: index.js 구조 수정
