+++ src/components/index.js
@@ -1,25 +1,14 @@
 // 컴포넌트 통합 내보내기
 
 // 공통 컴포넌트들
-export { default as Button } from './common/Button';
-export { default as Modal } from './common/Modal';
-export { default as Pagination } from './common/Pagination';
-export { default as ErrorBoundary } from './common/ErrorBoundary';
-export { default as LoadingIndicator } from './common/LoadingIndicator';
-export { default as ErrorIndicator } from './common/ErrorIndicator';
-export { default as DeleteConfirmModal } from './common/DeleteConfirmModal';
-
-// 지연 로딩 컴포넌트들
-export * from './common/LazyComponents';
+export * from './common';
 
 // 레이아웃 컴포넌트들
-export { default as Footer } from './layout/Footer';
-export { default as HeroSection } from './layout/HeroSection';
-export { default as NavigationHeader } from './layout/NavigationHeader';
+export * from './layout';
 
-// 리쿠르팅 컴포넌트들
-export { default as ApplicantFilters } from './recruiting/ApplicantFilters';
-export { default as ApplicantListSection } from './recruiting/ApplicantListSection';
-export { default as BulkStatusChangeModal } from './recruiting/BulkStatusChangeModal';
-export { default as RecruitingHeader } from './recruiting/RecruitingHeader';
-export { default as RecruitingInfoSection } from './recruiting/RecruitingInfoSection';
+// 리쿠르팅 컴포넌트들 (RecruitingDetail 페이지별 컴포넌트들)
+export { default as ApplicantFilters } from './pages/RecruitingDetail/filters/ApplicantFilters';
+export { default as ApplicantListSection } from './pages/RecruitingDetail/detail/ApplicantListSection';
+export { default as BulkStatusChangeModal } from './pages/RecruitingDetail/modals/BulkStatusChangeModal';
+export { default as RecruitingHeader } from './pages/RecruitingDetail/detail/RecruitingHeader';
+export { default as RecruitingInfoSection } from './pages/RecruitingDetail/detail/RecruitingInfoSection';
