+++ src/components/layout/index.js
@@ -1,5 +1,5 @@
 // Layout Components
 export { default as NavigationHeader } from './NavigationHeader';
-export { default as HeroSection } from './HeroSection/HeroSection';
-export { default as Footer } from './Footer/Footer';
-export { default as Header } from './Header/Header';
+export { default as HeroSection } from './HeroSection';
+export { default as Footer } from './Footer';
+export { default as Header } from './Header';
